requirements.txt
dj-database-url==0.5.0
Django==2.0.3
django-cors-headers==3.1.0
django-heroku==0.3.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2==2.8.3
PyJWT==1.7.1
python==3.6.8
pytz==2019.2
simplejson==3.16.0
whitenoise==4.1.3
